Rename index controller file
